+++ src/operations/remove.test.ts
@@ -1,6 +1,7 @@
-import { set, get } from '../operations'
+import { set, get, push } from '../operations'
 import { initializeApp, usersRef } from '../utilForTests'
 import { remove } from './remove'
+import { serverTimestamp } from '../fieldValue'
 
 initializeApp()
 describe('test remove', () => {
@@ -23,12 +24,12 @@ describe('test remove', () => {
 
 	it('test functionality', async () => {
 		const ref1 = usersRef('b/h/abc/m')
-		const ref2 = usersRef('b/d/k')
+		const ref2 = usersRef('b/h/abc/p')
 		await set(ref1, { abc: { n: '7' } })
 		await remove(ref1)
 		const data1 = (await get(ref1)).val()
 		expect(data1).toBe(null)
-		await set(ref2, 'hahaha')
+		await push(ref2, { r: serverTimestamp() })
 		await remove(ref2)
 		const data2 = (await get(ref2)).val()
 		expect(data2).toBe(null)